I found out that there already exists a Pippenger's implementation for [supranational/blst rust binding](https://github.com/supranational/blst/blob/a7fd1f584d26b0ae6cdc427976ea1d8980f7e15d/bindings/rust/src/pippenger.rs#L93-L238), which appears to use buckets (a large grid for storing the results) and a thread pool for parallelization.

Implementing it achieved the following results:

| Benchmark | Old | New | Speedup |
|--------|--------|--------|--------|
| blob_to_kzg_commitment | 54.145 ms | 17.132 ms | 3.16x |
| compute_kzg_proof | 65.029 ms | 26.509 ms | 2.45x |
| compute_blob_kzg_proof | 65.680 ms | 27.342 ms | 2.40x |
| bench_commit_to_poly | 348.61 ms | 97.402 ms | 3.58x | 

However, there seemed to be a little caveat that it didn't work for certain lengths:

* 0 --> panics with message `scalars length mismatch`
* 1 --> fails the `test_fk_multi_chunk_len_1_512` test
* 6 --> fails 1 out of 222 [test vectors](https://github.com/ethereum/c-kzg-4844/tree/main/tests/verify_blob_kzg_proof_batch/small)

So we opted out for a naive approach for lengths less than 8.

Related issue: #202